ajout sécurité avec les adresseurl incluant des id d'objects dans la db(pas accessible par d'autre utilisateurs)

+++ ASMAE/tennis/views.py
@@ -113,7 +113,7 @@ def inscriptionTournoi(request):
 			pair.extra1.add(ext)
 		
 		# Send mail
-		send_confirmation_email_pair_registered(Participant.objects.get(user=pair.user1), Participant.objects.get(user=pair.user2))
+		#send_confirmation_email_pair_registered(Participant.objects.get(user=pair.user1), Participant.objects.get(user=pair.user2))
 
 		pair.save()
 		return redirect(reverse(tournoi))
@@ -125,7 +125,12 @@ def inscriptionTournoi(request):
 	return redirect(reverse(home))
 
 def confirmPair(request,id):
+	pair = Pair.objects.filter(id=id)
+	if len(pair) <1:
+		return redirect(reverse(tournoi))
 	pair = Pair.objects.filter(id=id)[0]
+	if pair.user2 != request.user:
+		return redirect(reverse(tournoi))
 	if request.method == "POST":
 		if request.POST['action'] == "validate":
 			remarque = request.POST['remarque']
@@ -169,7 +174,12 @@ def confirmPair(request,id):
 	return redirect(reverse(home))
 
 def cancelPair(request,id):
+	pair = Pair.objects.filter(id=id)
+	if len(pair) <1:
+		return redirect(reverse(tournoi))
 	pair = Pair.objects.filter(id=id)[0]
+	if pair.user1 != request.user:
+		return redirect(reverse(tournoi))
 	if request.method == "POST":
 		#TODO check si il peut annuler cette pair
 		pair.delete()
@@ -191,9 +201,15 @@ def cancelPair(request,id):
 	return redirect(reverse(home))
 
 def viewPair(request,id):
+	pair = Pair.objects.filter(id=id)
+	if len(pair) <1:
+		return redirect(reverse(tournoi))
+	pair = Pair.objects.filter(id=id)[0]
+	if pair.user1 != request.user and pair.user2 != request.user:
+		return redirect(reverse(tournoi))
 	if request.user.is_authenticated():
 		#TODO check si il peut voir cette pair
-		pair = Pair.objects.filter(id=id)[0]
+
 		Ex = Extra.objects.all()
 		extra1 = pair.extra1.all()
 		extranot1 = list()
@@ -220,9 +236,14 @@ def viewPair(request,id):
 	return redirect(reverse(home))
 
 def payPair(request,id):
+	pair = Pair.objects.filter(id=id)
+	if len(pair) <1:
+		return redirect(reverse(tournoi))
+	pair = Pair.objects.filter(id=id)[0]
+	if pair.user1 != request.user and pair.user2 != request.user:
+		return redirect(reverse(tournoi))
 	if request.user.is_authenticated():
 		#TODO check si il peut payer cette pair
-		pair = Pair.objects.filter(id=id)[0]
 		Ex = Extra.objects.all()
 		extra1 = pair.extra1.all()
 		extra2 = pair.extra2.all()
@@ -282,7 +303,7 @@ def registerTerrain(request):
 		court = Court(rue = rue,numero=numero,boite=boite,codepostal=postalcode,localite=locality,acces=acces,matiere=matiere,type=type,dispoDimanche=dispoDimanche,dispoSamedi=dispoSamedi,etat= etat,commentaire=commentaire,user = request.user)
 	
 		# Send confirmation mail
-		send_confirmation_email_court_registered(Participant.objects.get(user=request.user), court)
+		##send_confirmation_email_court_registered(Participant.objects.get(user=request.user), court)
 		
 		court.save()
 		
@@ -293,7 +314,14 @@ def registerTerrain(request):
 	return redirect(reverse(home))
 
 def editTerrain(request,id):
+	court = Court.objects.filter(id=id)
+
+	if len(court) <1:
+		return redirect(reverse(terrain))
 	court = Court.objects.filter(id=id)[0]
+	if court.user != request.user:
+		return redirect(reverse(terrain))
+
 	if request.method == "POST":
 		if request.POST['action'] == "modifyCourt":
 			rue = request.POST['rue']
@@ -388,59 +416,58 @@ def staffPaire(request):
 	return redirect(reverse(home))
 
 def staffExtra(request):
-	#List of Extra
-	Ex = Extra.objects.all()
-	if request.method == "POST":
-		if request.POST['action'] == "addExtra":
-			nom = request.POST['name']
-			prix = request.POST['price']
-			message = request.POST['message']
+	if request.user.is_staff: #TODO
+		Ex = Extra.objects.all()
+		if request.method == "POST":
+			if request.POST['action'] == "addExtra":
+				nom = request.POST['name']
+				prix = request.POST['price']
+				message = request.POST['message']
 			
-			if nom=="":
-				errorAdd = "Veuillez rajouter un nom à l'extra!"
-				return render(request,'tennis/staffExtra.html',locals())			
+				if nom=="":
+					errorAdd = "Veuillez rajouter un nom à l'extra!"
+					return render(request,'tennis/staffExtra.html',locals())			
 
-			if not is_number(prix):
-				errorAdd = "Le prix n'a pas le bon format"
-				return render(request,'tennis/staffExtra.html',locals())
+				if not is_number(prix):
+					errorAdd = "Le prix n'a pas le bon format"
+					return render(request,'tennis/staffExtra.html',locals())
 			
-			extra = Extra(nom=nom,prix=prix,commentaires = message)
-			extra.save()
+				extra = Extra(nom=nom,prix=prix,commentaires = message)
+				extra.save()
 
-			successAdd = "Extra bien ajouté!"
+				successAdd = "Extra bien ajouté!"
 
-		if request.POST['action'] == "modifyExtra":
-			id = request.POST['id']
-			nom = request.POST['name']
-			prix = request.POST['price']
-			message = request.POST['message']
+			if request.POST['action'] == "modifyExtra":
+				id = request.POST['id']
+				nom = request.POST['name']
+				prix = request.POST['price']
+				message = request.POST['message']
 			
-			extra = Extra.objects.filter(id = id)[0]
+				extra = Extra.objects.filter(id = id)[0]
 	
-			if nom=="":
-				errorEdit = "Veuillez rajouter un nom à l'extra!"
-				return render(request,'tennis/staffExtra.html',locals())			
+				if nom=="":
+					errorEdit = "Veuillez rajouter un nom à l'extra!"
+					return render(request,'tennis/staffExtra.html',locals())			
 
-			if not is_number(prix):
-				errorEdit = "Le prix n'a pas le bon format"
-				return render(request,'tennis/staffExtra.html',locals())	
+				if not is_number(prix):
+					errorEdit = "Le prix n'a pas le bon format"
+					return render(request,'tennis/staffExtra.html',locals())	
 			
 			
-			extra.nom = nom
-			extra.prix = prix
-			extra.commentaires = message
-			extra.save()
-			successEdit = "Extra bien édité!"
+				extra.nom = nom
+				extra.prix = prix
+				extra.commentaires = message
+				extra.save()
+				successEdit = "Extra bien édité!"
 
-		if request.POST['action'] == "deleteExtra":
-			id = request.POST['id']
-			extra = Extra.objects.filter(id = id)[0]
-			extra.delete()
-			successDelete = "Extra bien supprimé!"
+			if request.POST['action'] == "deleteExtra":
+				id = request.POST['id']
+				extra = Extra.objects.filter(id = id)[0]
+				extra.delete()
+				successDelete = "Extra bien supprimé!"
 
-	if request.user.is_authenticated():
-		if request.user.is_staff: #TODO
-		    return render(request,'tennis/staffExtra.html',locals())
+		if request.user.is_authenticated():
+			return render(request,'tennis/staffExtra.html',locals())
 	return redirect(reverse(home))
 
 def staffUser(request):
@@ -451,8 +478,16 @@ def staffUser(request):
 	return redirect(reverse(home))
 
 def viewUser(request,name):
-	#TODO check si l'user use exist sinon retourner une page 404
-	use = User.objects.filter(username=name)[0]
+
+	use = User.objects.filter(username=name)
+	
+
+	if len(use) <1:
+		return redirect(reverse(staffUser))
+	use = Court.objects.filter(id=id)[0]
+	if use != request.user:
+		return redirect(reverse(staffUser))
+
 	birthdate = use.participant.datenaissance
 	formatedBirthdate = birthdate.strftime('%d/%m/%Y')
 	terrain = Court.objects.filter(user=use)
@@ -465,132 +500,131 @@ def viewUser(request,name):
 	return redirect(reverse(home))
 
 def validateTerrain(request, id):
-	court = Court.objects.filter(id=id)[0]
-	if request.method == "POST":
-		message = request.POST['message']
-		if request.POST.__contains__("valide"):
-			valide = True
-		else:
-			valide = False
+	if request.user.is_staff:
+		court = Court.objects.filter(id=id)[0]
+		if request.method == "POST":
+			message = request.POST['message']
+			if request.POST.__contains__("valide"):
+				valide = True
+			else:
+				valide = False
 		
-		court.commentaireStaff = message
-		court.valide = valide
-		court.save()
-		successEdit = "Terrain bien édité!"
+			court.commentaireStaff = message
+			court.valide = valide
+			court.save()
+			successEdit = "Terrain bien édité!"
 		
-	if request.user.is_authenticated():
-		if request.user.is_staff:
-		
+		if request.user.is_authenticated():		
 			return render(request,'tennis/validateTerrain.html',locals())
 	return redirect(reverse(home))
 
 def editTerrainStaff(request, id):
-	court = Court.objects.filter(id=id)[0]
-	if request.method == "POST":
-		if request.POST['action'] == "modifyCourt":
-			rue = request.POST['rue']
-			numero = request.POST['numero']
-			boite = request.POST['boite']
-			postalcode = request.POST['postalcode']	
-			locality = request.POST['loclity']
-			acces = request.POST['acces']
-			matiere = request.POST['matiere']
-			type = request.POST['type']
-			etat = request.POST['etat']
-			commentaire = request.POST['comment']
-			if request.POST.__contains__("dispoSamedi"):
-					dispoSamedi = True
-			else:
-				dispoSamedi = False
-			if request.POST.__contains__("dispoDimanche"):
-					dispoDimanche = True
-			else:
-				dispoDimanche = False
+	if request.user.is_staff:
+		court = Court.objects.filter(id=id)[0]
+		if request.method == "POST":
+			if request.POST['action'] == "modifyCourt":
+				rue = request.POST['rue']
+				numero = request.POST['numero']
+				boite = request.POST['boite']
+				postalcode = request.POST['postalcode']	
+				locality = request.POST['loclity']
+				acces = request.POST['acces']
+				matiere = request.POST['matiere']
+				type = request.POST['type']
+				etat = request.POST['etat']
+				commentaire = request.POST['comment']
+				if request.POST.__contains__("dispoSamedi"):
+						dispoSamedi = True
+				else:
+					dispoSamedi = False
+				if request.POST.__contains__("dispoDimanche"):
+						dispoDimanche = True
+				else:
+					dispoDimanche = False
 
 
-			if (rue=="" or numero=="" or postalcode=="" or locality=="" or matiere=="" or type=="" or etat==""):
-				errorAdd = "Veuillez remplir tous les champs obligatoires !"
-				return render(request,'tennis/registerTerrain.html',locals())
+				if (rue=="" or numero=="" or postalcode=="" or locality=="" or matiere=="" or type=="" or etat==""):
+					errorAdd = "Veuillez remplir tous les champs obligatoires !"
+					return render(request,'tennis/registerTerrain.html',locals())
 
 		
-			court.rue = rue 		
-			court.numero=numero
-			court.boite=boite
-			court.codepostal=postalcode
-			court.localite=locality
-			court.acces=acces
-			court.matiere=matiere
-			court.type=type
-			court.dispoDimanche=dispoDimanche
-			court.dispoSamedi=dispoSamedi
-			court.etat= etat
-			court.commentaire=commentaire
-			court.user = request.user
-			court.save()
-			#successEdit = "Terrain "+str(id)+" bien édité!"
-			return redirect(reverse(validateTerrain,args={id}))
+				court.rue = rue 		
+				court.numero=numero
+				court.boite=boite
+				court.codepostal=postalcode
+				court.localite=locality
+				court.acces=acces
+				court.matiere=matiere
+				court.type=type
+				court.dispoDimanche=dispoDimanche
+				court.dispoSamedi=dispoSamedi
+				court.etat= etat
+				court.commentaire=commentaire
+				court.user = request.user
+				court.save()
+				#successEdit = "Terrain "+str(id)+" bien édité!"
+				return redirect(reverse(validateTerrain,args={id}))
 
-		if request.POST['action'] == "deleteCourt":
-			#TODO delete terrain staff
-			court.delete()
-			return redirect(reverse(staffTerrain))
-	if request.user.is_authenticated():
-		if request.user.is_staff:
+			if request.POST['action'] == "deleteCourt":
+				#TODO delete terrain staff
+				court.delete()
+				return redirect(reverse(staffTerrain))
+		if request.user.is_authenticated():	
 			return render(request,'tennis/editTerrainStaff.html',locals())
 	return redirect(reverse(home))
 
 def validatePair(request, id):
-	pair = Pair.objects.filter(id=id)[0]
-	if request.method == "POST":
-		if request.POST['action'] == "editPair":
-			valid = request.POST['valid']
-			paid = request.POST['pay']
-			if valid == "Oui":
-				valider = True
-			else:
-				valider = False
-			if paid == "Oui":
-				payer = True
-			else:
-				payer = False
-			pair.valid = valider
-			pair.pay = payer
-			pair.save()
+	if request.user.is_staff:
+		pair = Pair.objects.filter(id=id)[0]
+		if request.method == "POST":
+			if request.POST['action'] == "editPair":
+				valid = request.POST['valid']
+				paid = request.POST['pay']
+				if valid == "Oui":
+					valider = True
+				else:
+					valider = False
+				if paid == "Oui":
+					payer = True
+				else:
+					payer = False
+				pair.valid = valider
+				pair.pay = payer
+				pair.save()
 
-			return redirect(reverse(staffPaire))
+				return redirect(reverse(staffPaire))
 
-		if request.POST['action'] == "deletePair":
-			pair.delete()
-			return redirect(reverse(staffPaire))
+			if request.POST['action'] == "deletePair":
+				pair.delete()
+				return redirect(reverse(staffPaire))
 			
 	
-	Ex = Extra.objects.all()
-	extra1 = pair.extra1.all()
-	extranot1 = list()
-	for elem in Ex:
-		contained = False
-		for el in extra1:
-			if elem.id == el.id:
-				contained = True
-		if contained == False:	
-			extranot1.append(Extra.objects.filter(id=elem.id)[0])
+		Ex = Extra.objects.all()
+		extra1 = pair.extra1.all()
+		extranot1 = list()
+		for elem in Ex:
+			contained = False
+			for el in extra1:
+				if elem.id == el.id:
+					contained = True
+			if contained == False:	
+				extranot1.append(Extra.objects.filter(id=elem.id)[0])
 
-	extra2 = pair.extra2.all()
-	extranot2 = list()
-	for elem in Ex:
-		contained = False
-		for el in extra2:
-			if elem.id == el.id:
-				contained = True
-		if contained == False:	
-			extranot2.append(Extra.objects.filter(id=elem.id)[0])
+		extra2 = pair.extra2.all()
+		extranot2 = list()
+		for elem in Ex:
+			contained = False
+			for el in extra2:
+				if elem.id == el.id:
+					contained = True
+			if contained == False:	
+				extranot2.append(Extra.objects.filter(id=elem.id)[0])
 
-	birthdate1 = pair.user1.participant.datenaissance
-	formatedBirthdate1 = birthdate1.strftime('%d/%m/%Y')
-	birthdate2 = pair.user2.participant.datenaissance
-	formatedBirthdate2 = birthdate2.strftime('%d/%m/%Y')
-	if request.user.is_authenticated():
-		if request.user.is_staff:
+		birthdate1 = pair.user1.participant.datenaissance
+		formatedBirthdate1 = birthdate1.strftime('%d/%m/%Y')
+		birthdate2 = pair.user2.participant.datenaissance
+		formatedBirthdate2 = birthdate2.strftime('%d/%m/%Y')
+		if request.user.is_authenticated():
 			return render(request,'tennis/validatePair.html',locals())
 	return redirect(reverse(home))
